Conflicts:
	DoongsilDoongsil/pom.xml
	DoongsilDoongsil/src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
	DoongsilDoongsil/src/main/webapp/resources/mybatis/mappers/paboard.xml
	DoongsilDoongsil/target/m2e-wtp/web-resources/META-INF/maven/doongsil.com/web/pom.xml

+++ DoongsilDoongsil/src/main/java/doongsil/com/web/homework/controller/HomeworkController.java
@@ -3,7 +3,10 @@ package doongsil.com.web.homework.controller;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.sql.Date;
 import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -176,6 +179,21 @@ public class HomeworkController {
 			HashMap<Integer, String[]> type2contents = service.makeType2Visible(data.getTho_contents());
 			System.out.println(type2contents);
 			mod.addObject("type2contents", type2contents);
+		} else if(data.getTho_homeworktype() == 3) {
+			java.sql.Date startDate = data.getTho_writedate();
+			java.sql.Date endDate = data.getTho_deadline();
+			LocalDate startLocal = startDate.toLocalDate();
+			
+			long endMinusStartSec = (endDate.getTime() - startDate.getTime()) / 1000;
+			long duration = endMinusStartSec / (24*60*60);
+			int durDate = Long.valueOf(duration).intValue();
+			//mod.addObject("type3duration", durDate);
+			
+			LocalDate[] type3DayArr = new LocalDate[durDate + 1];
+			for(int type3i = 0; type3i <= durDate; type3i++) {
+				type3DayArr[type3i] = startLocal.plusDays(type3i);
+			}
+			mod.addObject("type3DayArr", type3DayArr);
 		}
 		
 		//글 값 넣기
@@ -187,6 +205,14 @@ public class HomeworkController {
 		List<S_HomeworkVO> sdatas = service.selectSHList(vo.getTho_id());
 		if(sdatas != null) {
 			mod.addObject("sdatas", sdatas);
+			HashMap<Integer, String[]> studentImages = new HashMap<Integer, String[]>();
+			for(S_HomeworkVO simg: sdatas) {
+				String[] simgarrs = new String[1];
+				if(simg.getSho_fileurl() != null) {
+					simgarrs = service.getImgList(simg.getSho_fileurl());
+					studentImages.put(simg.getSho_id(), simgarrs);
+				}
+			}
 			if(data.getTho_homeworktype() == 2) {
 				HashMap<Integer, String[]> type2answers = new HashMap<Integer, String[]>();
 				HashMap<Integer, String[]> type2teacGBs = new HashMap<Integer, String[]>();
@@ -198,6 +224,7 @@ public class HomeworkController {
 						type2teacGBs.put(s.getSho_id(), teacGBarr);
 					}
 				}
+				mod.addObject("studentImages", studentImages);
 				mod.addObject("type2answers", type2answers);
 				mod.addObject("type2GBforTeacher", type2teacGBs);
 			}
@@ -209,7 +236,7 @@ public class HomeworkController {
 		//만약 학생일 경우
 		S_HomeworkVO shwVO = new S_HomeworkVO();
 		shwVO.setSho_tid(data.getTho_id()); 
-		shwVO.setSho_writer(4);//session에서 값 받아서 설정하도록 나중에 수정
+		shwVO.setSho_writer(7);//session에서 값 받아서 설정하도록 나중에 수정
 		List<S_HomeworkVO> sworks = service.selectStudentHWs(shwVO);
 		if(sworks == null || sworks.size() == 0) {
 			mod.addObject("sworksnull", true);
@@ -217,7 +244,20 @@ public class HomeworkController {
 		} else {
 			mod.addObject("sworksnull", false);
 			System.out.println("학생 숙제값 있음");
-			mod.addObject("sworks", sworks);
+			if(data.getTho_homeworktype() != 3) {
+				mod.addObject("sworks", sworks);
+			}
+			
+			HashMap<Integer, String[]> studentImagesForS = new HashMap<Integer, String[]>();
+			for(S_HomeworkVO simgForS: sworks) {
+				String[] simgarrsForS = new String[1];
+				if(simgForS.getSho_fileurl() != null) {
+					simgarrsForS = service.getImgList(simgForS.getSho_fileurl());
+					studentImagesForS.put(simgForS.getSho_id(), simgarrsForS);
+				}
+			}
+			mod.addObject("studentImagesForS", studentImagesForS);
+			
 			if(data.getTho_homeworktype() == 2) {
 				String[] ansarr = service.makeType2Answer(sworks.get(0));
 				mod.addObject("sanswer", ansarr);
@@ -226,9 +266,22 @@ public class HomeworkController {
 					String[] GBarr = sworks.get(0).getSho_goodbad().split("");
 					mod.addObject("type2GBforStudent", GBarr);
 				}
+			} else if(data.getTho_homeworktype() == 3) {
+				HashMap<java.sql.Date, S_HomeworkVO> type3sWorks = new HashMap<java.sql.Date, S_HomeworkVO>();
+				HashMap<java.sql.Date, String[]> type3sWeather = new HashMap<java.sql.Date, String[]>();
+				HashMap<java.sql.Date, String> type3comment = new HashMap<java.sql.Date, String>();
+				for(S_HomeworkVO stype3: sworks) {
+					String[] temps3arr = stype3.getSho_contents().split("\\|\\|");
+					java.sql.Date temp3sdate = stype3.getSho_date();
+					type3sWorks.put(temp3sdate, stype3);
+					type3sWeather.put(temp3sdate, temps3arr);
+					type3comment.put(temp3sdate, stype3.getSho_comment() != null ? "true" : "false");
+				}
+				mod.addObject("type3sWorks", type3sWorks);
+				mod.addObject("type3sWC", type3sWeather);
 			}
 		}
-		mod.setViewName("homework/detail");
+		mod.setViewName("homework/detail/detail");
 		return mod;
 		
 	}
@@ -304,13 +357,32 @@ public class HomeworkController {
 		return "homework/empty";
 	}
 	
+	@RequestMapping(value="/homework/type3", method=RequestMethod.GET)
+	public String homeworktest3(@RequestParam("date") java.sql.Date sho_date) {
+		
+		return "homework/detail/type3";
+	}
+	
 	@RequestMapping(value="/deleteSH", method=RequestMethod.GET)
 	public ModelAndView deleteSH(ModelAndView mod, @RequestParam("sho_id") int sho_id) {
+		boolean res = false;
+		
+		//삭제할 vo객체 생성
 		S_HomeworkVO vo = new S_HomeworkVO();
 		vo.setSho_id(sho_id);
-		boolean res = service.deleteSH(vo);
-		if(res) {
-			mod.setViewName("redirect:/homework");
+		
+		//채점이 되어있는지 확인
+		S_HomeworkVO data = service.selectOneSH(sho_id);
+		if(data != null) {
+			if(data.getSho_comment() == null && data.getSho_goodbad() == null) {
+				res = service.deleteSH(vo);
+			}
+			
+			if(res) {
+				mod.setViewName("redirect:/homework");
+			} else {
+				mod.setViewName("homework/error");
+			}
 		} else {
 			mod.setViewName("homework/error");
 		}
@@ -319,13 +391,17 @@ public class HomeworkController {
 	
 	@RequestMapping(value="/deleteHW", method=RequestMethod.GET)
 	public ModelAndView deleteHW(ModelAndView mod, @RequestParam("tho_id") int tho_id) {
+		boolean res = false;
+		
 		T_HomeworkVO vo = new T_HomeworkVO();
 		vo.setTho_id(tho_id);
-		boolean res = service.deleteHW(vo);
-		//해당 id를 tid로 가지고 있는 숙제까지 삭제하는 로직 짤것
+		
+		List<S_HomeworkVO> datas = service.selectSHList(tho_id);
+		if(datas.size() <= 0) {
+			res = service.deleteHW(vo);
+		}
+		
 		if(res) {
-			//mod.setViewName("homework/list");
-			//return new ModelAndView("homework/list");
 			mod.setViewName("redirect:/homework");
 		} else {
 			mod.setViewName("homework/error");
